refactor(contact-center): consolidate wxApp visibility helpers in cc-components
